python/requirements.txt
attrs==23.1.0
mysql-connector-python==8.1.0
numpy==1.26.1
protobuf==4.21.12
